move reportViewedDashboard to eventUsageLogic
